clean up

+++ pico_platformio_touch_mouse/src/main.cpp
@@ -1,30 +1,38 @@
-#include "config.h"
+/*
+ * main.cpp — XPT2046 resistive touch to USB HID Mouse
+ *
+ * PlatformIO / Arduino-Pico (earlephilhower) core.
+ * The SPI peripheral and default pin mapping are configured in platformio.ini;
+ * only CS and the PENIRQ interrupt line need explicit pin management here.
+ */
 
+#include "config.h"
 #include <Mouse.h>
 
-// SPI pins
-#define PIN_CS 17
-#define PIN_INTR 20
-#define LED_PIN 25
+// ── Hardware pin assignments ───────────────────────────────────────────────
+// SPI0 SCK/MOSI/MISO are set via build flags in platformio.ini.
+#define PIN_CS   17   // XPT2046 chip select, active low
+#define PIN_INTR 20   // XPT2046 PENIRQ, active low (per datasheet)
+#define LED_PIN  25   // Onboard LED, PWM-controlled
 
-// Touch controller commands for XPT2046
-#define CMD_READ_X 0xD0
-#define CMD_READ_Y 0x90
+// ── XPT2046 command bytes ─────────────────────────────────────────────────
+#define CMD_READ_X 0xD0   // Differential X measurement, 12-bit
+#define CMD_READ_Y 0x90   // Differential Y measurement, 12-bit
 
-// Screen resolution (adjust as needed)
-#define SCREEN_WIDTH 800
+// ── Screen mapping ────────────────────────────────────────────────────────
+#define SCREEN_WIDTH  800
 #define SCREEN_HEIGHT 480
 
-// Touch raw value ranges
-#define X_MIN 150
-#define X_MAX 3784
-#define Y_MIN 277
-#define Y_MAX 3784
+// ── Touch calibration (raw 12-bit ADC limits, measured on panel) ──────────
+#define X_MIN   150
+#define X_MAX  3784
+#define Y_MIN   277
+#define Y_MAX  3784
 #define X_RANGE (X_MAX - X_MIN)
 #define Y_RANGE (Y_MAX - Y_MIN)
 
-// Mouse sensitivity
-#define MOUSE_SCALE 1
+// ── Tuning ────────────────────────────────────────────────────────────────
+#define MOUSE_SCALE 1   // Multiplier on movement deltas; 1 = 1:1
 
 static uint16_t read_touch_axis(uint8_t cmd) {
     SPI.beginTransaction(SPISettings(1000000, MSBFIRST, SPI_MODE0));
@@ -47,8 +55,8 @@ void setup() {
     pinMode(PIN_CS, OUTPUT);
     digitalWrite(PIN_CS, HIGH);
 
-    // Interrupt pin
-    pinMode(PIN_INTR, INPUT_PULLUP); // Assuming active low
+    // PENIRQ is open-drain, active low — internal pull-up required
+    pinMode(PIN_INTR, INPUT_PULLUP);
 
     // LED pin
     pinMode(LED_PIN, OUTPUT);
@@ -74,12 +82,12 @@ void loop() {
             analogWrite(LED_PIN, 0); // Turn off LED on touch start
         } else {
             unsigned long elapsed = millis() - touch_start;
-            int brightness = 255 - (elapsed / 10); 
+            int brightness = 255 - (int)(elapsed / 10);
             if (brightness < 0) brightness = 0;
             analogWrite(LED_PIN, brightness);
         }
 
-        // Average 2 readings to reduce noise
+        // Step 1 — Average 2 raw ADC readings per axis to reduce noise
         uint16_t x_raw = 0, y_raw = 0;
         for (int i = 0; i < 2; i++) {
             x_raw += read_touch_axis(CMD_READ_X);
@@ -88,44 +96,45 @@ void loop() {
         x_raw /= 2;
         y_raw /= 2;
 
-        // Print raw values over UART
+        // Debug output on UART0 (GPIO 0 TX, 115200 baud)
         Serial1.print("X: ");
         Serial1.print(x_raw);
         Serial1.print(", Y: ");
         Serial1.println(y_raw);
 
-        // Map raw values to screen coordinates with offsets
-        int16_t x = ((x_raw - X_MIN) * SCREEN_WIDTH) / X_RANGE;
+        // Step 2 — Map raw ADC values to screen pixel coordinates
+        int16_t x = ((x_raw - X_MIN) * SCREEN_WIDTH)  / X_RANGE;
         int16_t y = ((y_raw - Y_MIN) * SCREEN_HEIGHT) / Y_RANGE;
 
         // Clamp to screen bounds
-        x = constrain(x, 0, SCREEN_WIDTH - 1);
+        x = constrain(x, 0, SCREEN_WIDTH  - 1);
         y = constrain(y, 0, SCREEN_HEIGHT - 1);
 
         if (last_x != -1 && last_y != -1) {
             int8_t dx = (x - last_x) * MOUSE_SCALE;
             int8_t dy = (y - last_y) * MOUSE_SCALE;
 
-            // Only move if change is significant to avoid jitter
+            // Step 3 — Jitter suppression: discard sub-threshold deltas
             if (abs(dx) >= 10 || abs(dy) >= 10) {
                 Mouse.move(dx, dy);
                 last_x = x;
                 last_y = y;
             }
         } else {
-            // Initialize last position to center of screen for absolute-like jump on first touch
-            last_x = SCREEN_WIDTH / 2;
+            // Seed from screen centre on first touch so the first delta is
+            // relative to a known position rather than an uninitialised one
+            last_x = SCREEN_WIDTH  / 2;
             last_y = SCREEN_HEIGHT / 2;
         }
 
-        delay(15); // Debounce
+        delay(15);
     } else {
         if (touching) {
             touching = false;
             last_touch_time = millis();
-            analogWrite(LED_PIN, 255); // Back to full brightness when touch ends
+            analogWrite(LED_PIN, 255);
         }
-        // Reset position if no touch for 300 milliseconds
+        // Reset tracking state 300 ms after lift-off
         if (millis() - last_touch_time > 300) {
             last_x = -1;
             last_y = -1;
